Refactor the logger into a few different files
