+++ packages/server/src/lib/logging/base.ts
@@ -1,0 +1,26 @@
+import type winston from "winston";
+
+export const SyslogLevels = {
+  emerg: 0,
+  alert: 1,
+  crit: 2,
+  error: 3,
+  warning: 4,
+  notice: 5,
+  info: 6,
+  debug: 7,
+  trace: 8,
+} as const satisfies winston.config.AbstractConfigSetLevels;
+export type SyslogLevels = keyof typeof SyslogLevels;
+
+export const syslogColors = {
+  emerg: "red",
+  alert: "yellow",
+  crit: "red",
+  error: "red",
+  warning: "red",
+  notice: "yellow",
+  info: "green",
+  debug: "blue",
+  trace: "gray",
+} satisfies winston.config.AbstractConfigSetColors;
