+++ packages/server/src/lib/logging/logger.ts
@@ -1,0 +1,2 @@
+export { sqlLogger } from "./sqlLogging.js";
+export { logFatal, logger } from "./standardLogging.js";
