+++ packages/server/src/server.ts
@@ -10,7 +10,7 @@ import Koa from "koa";
 import { koaBody } from "koa-body";
 
 import { applicationHost, applicationPort } from "./environment.js";
-import { logger } from "./logger.js";
+import { logger } from "./lib/logging/logger.js";
 import type { GraphQLContext } from "./resolvers/context.js";
 import eventsApiRouter from "./routes/api/events/index.js";
 import healthCheckRouter from "./routes/api/healthcheck/index.js";
