+++ packages/server/src/umzug.ts
@@ -5,7 +5,7 @@ import { fileURLToPath, pathToFileURL } from "node:url";
 import { DataTypes } from "@sequelize/core";
 import type { RunnableMigration } from "umzug";
 
-import { logger } from "./logger.js";
+import { logger } from "./lib/logging/logger.js";
 import type { MigrationContext } from "./migrations/migrationContext.js";
 const { sequelizeDb } = await import("./data-source.js");
 const { Umzug, SequelizeStorage } = await import("umzug");
